requirements.txt
dj-database-url==0.5.0
Django==2.1.7
django-cors-headers==2.5.2
django-default-imagefield==0.5.5
django-heroku==0.3.1
djangorestframework==3.9.2
drf-extensions==0.4.0
gunicorn==19.9.0
Pillow==5.4.1
postgres==2.2.2
psycopg2==2.7.7
psycopg2-binary==2.7.7
pytz==2018.9
whitenoise==4.1.2
